Renamed item and player pages
Renamed these in anticipation of having an item.html that would have the details of the item via a json file and display without going to WoWHead. We also need a player page that has a list of active guild members so that we can determine those awarded loot / dkp no longer on the list.

Added highcharity.css as a start to be the stylesheet to make it look pretty. Lots of commits ahead.